requirements.txt
aiogram==3.4.0
Flask==2.2.3
python-dotenv==1.0.0